# Speedup graph traversal functions