--- a/doc/gallery/optimize/root.ipynb
+++ b/doc/gallery/optimize/root.ipynb
@@ -1772,7 +1772,9 @@
     }
    ],
    "source": [
-    "from pytensor.graph.basic import explicit_graph_inputs\n",
+    "\n",
+    "from pytensor.graph.traversal import explicit_graph_inputs\n",
+    "\n",
     "list(explicit_graph_inputs(w_bar_2))"
    ]
   },