--- a/pytensor/compile/function/types.py
+++ b/pytensor/compile/function/types.py
@@ -20,14 +20,13 @@ from pytensor.configdefaults import config
 from pytensor.graph.basic import (
     Constant,
     Variable,
-    ancestors,
     clone_get_equiv,
-    graph_inputs,
 )
 from pytensor.graph.destroyhandler import DestroyHandler
 from pytensor.graph.features import AlreadyThere, Feature, PreserveVariableAttributes
 from pytensor.graph.fg import FunctionGraph
 from pytensor.graph.op import HasInnerGraph
+from pytensor.graph.traversal import ancestors, graph_inputs
 from pytensor.graph.utils import InconsistencyError, get_variable_trace_string
 from pytensor.link.basic import Container
 from pytensor.link.utils import raise_with_op