… pytensor/graph/traversal.py
… pytensor/printing.py
… pytensor/scalar/basic.py
… pytensor/scan/basic.py
… pytensor/scan/op.py
… pytensor/scan/rewriting.py
… pytensor/scan/utils.py
… pytensor/tensor/_linalg/solve/rewriting.py
… pytensor/tensor/blas.py
… pytensor/tensor/blockwise.py
… pytensor/tensor/conv/abstract_conv.py
… pytensor/tensor/optimize.py
… pytensor/tensor/random/rewriting/basic.py
… pytensor/tensor/random/rewriting/jax.py
… pytensor/tensor/random/rewriting/numba.py
… pytensor/tensor/rewriting/basic.py
… pytensor/tensor/rewriting/blas.py
… pytensor/tensor/rewriting/blas_c.py
… pytensor/tensor/rewriting/blockwise.py
… pytensor/tensor/rewriting/elemwise.py
… pytensor/tensor/rewriting/jax.py
… pytensor/tensor/rewriting/linalg.py
… pytensor/tensor/rewriting/numba.py
… pytensor/tensor/rewriting/ofg.py
… pytensor/tensor/rewriting/shape.py
… pytensor/tensor/rewriting/subtensor.py
… pytensor/xtensor/rewriting/utils.py
… tests/compile/test_profiling.py
… tests/graph/rewriting/test_kanren.py
… tests/graph/test_basic.py